+++ resources/[dg]/dg-admin/typescript/src/server/modules/commands/data/cloak.ts
@@ -8,7 +8,7 @@ export const cloak: CommandData = {
   isClientCommand: false,
   target: [],
   role: 'support',
-  handler: caller => {
+  handler: async caller => {
     const toggle = !getPlayerCommandState(caller.source, 'cloak');
     setPlayerCommandState(caller.source, 'cloak', toggle);
     Events.emitNet('admin:commands:cloak', caller.source, toggle);
@@ -16,10 +16,18 @@ export const cloak: CommandData = {
     // do some logic because we want to stay invis when exiting cloak but in noclip
     const inNoclip = getPlayerCommandState(caller.source, 'noclip');
     const visible = toggle ? false : !inNoclip;
-    Sync.setPlayerVisible(caller.source, visible);
 
-    Notifications.add(caller.source, `Cloak ${toggle ? 'enabled' : 'disabled'}`);
-    allowInvisibleForInteractingPlayers(caller.source, toggle);
+    // when entering cloak, we FIRST allow interacint players before actually toggling
+    if (toggle) {
+      await allowInvisibleForInteractingPlayers(caller.source, true);
+    }
+
+    await Sync.setPlayerVisible(caller.source, visible);
+
+    // when exitting cloak, we FIRST exit before disallowing interacting players
+    if (!toggle) {
+      await allowInvisibleForInteractingPlayers(caller.source, false);
+    }
   },
   UI: {
     title: 'Cloak',